requirements.txt
tensorflow==2.13.0
Pillow
numpy
matplotlib
icrawler